waitm-637: multistep form refactor
